+++ week-7/assignment-1/index.ts
@@ -1,0 +1,30 @@
+
+import express  from "express";
+import mongoose from "mongoose";
+
+const app = express();
+const port = 3000;
+import authRoutes from "./routes/auth";
+import todoRoutes from "./routes/todo";
+import cors from "cors";
+
+app.use(cors());
+app.use(express.json());
+app.use("/auth", authRoutes);
+app.use("/todo", todoRoutes);
+
+app.listen(port, () => {
+    console.log(`Example app listening at http://localhost:${port}`)
+})
+
+mongoose.connect('mongodb+srv://admin:admin@cluster0.y6dverl.mongodb.net/', { useNewUrlParser: true, useUnifiedTopology: true });
+
+const db = mongoose.connection;
+
+db.on("error", (error) => {
+  console.error("MongoDB connection error:", error);
+});
+
+db.once("connected", () => {
+  console.log("db connected");
+});
