+++ week-7/assignment-1/middleware/index.ts
@@ -1,8 +1,7 @@
-const jwt = require('jsonwebtoken');
-const { Response } = require('express');
-const SECRET = 'SECr3t';  // This should be in an environment variable in a real application
+import jwt from 'jsonwebtoken';
+export const SECRET = 'SECr3t'; // This should be in an environment variable in a real application
 
-const authenticateJwt = (req, res, next) => {
+export const authenticateJwt = (req, res, next) => {
   const authHeader = req.headers.authorization;
   if (authHeader) {
     const token = authHeader.split(' ')[1];
@@ -17,8 +16,3 @@ const authenticateJwt = (req, res, next) => {
     res.sendStatus(401);
   }
 };
-
-module.exports = {
-    authenticateJwt,
-    SECRET
-}
