+++ week-7/assignment-1/package.json
@@ -4,6 +4,7 @@
   "description": "Convert this from JS to typescript.",
   "main": "index.js",
   "scripts": {
+    "start": "nodemon index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -13,6 +14,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.1",
-    "mongoose": "^7.4.1"
+    "mongoose": "^7.4.1",
+    "nodemon": "^3.0.1"
   }
 }
